Refactor runtime
